refactor(style): split style.ts into a style/ subdirectory
style.ts was the outlier among transformers (~830 lines vs 70-160 for the
others). Split the helpers into a style/ subdirectory, mirroring the existing
layout.ts + layout/ pattern so style.ts stays the single public entry point:

- style/color.ts   — hex/rgba conversion + solid-fill flatten compositing
- style/image.ts   — scale-mode translation, transform hashing, pattern fills
- style/gradient.ts — gradient geometry (the biggest chunk, ~290 lines)

style.ts keeps the cross-cutting pieces (SimplifiedFill/SimplifiedStroke,
parsePaint dispatch, buildSimplifiedStrokes) and re-exports the submodules'
public surface, so every caller's '~/transformers/style.js' import is unchanged.

Pure structural move, no behavior change. The one type tweak: flattenSolidFills
is now typed to its actual return (CSSHexColor | CSSRGBAColor | null) instead of
the wider SimplifiedFill, since it only ever yields a color.
